requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
reportlab==4.0.8
python-dotenv==1.0.0
gunicorn==21.2.0
pypdf==3.17.0
python-docx==1.1.0
requests==2.31.0
